+++ BackEnd/AgendaTypeScript/template-ts/package.json
@@ -4,8 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "ts-node ./src/index.ts",
-    "build": "tsc && node ./build/index.js"
+    "start": "ts-node ./src/exercicios.ts",
+    "build": "tsc && node ./build/exercicios.js"
   },
   "author": "",
   "license": "ISC",
