--- BackEnd/AgendaTypeScript/template-ts/src/index.js
@@ -1,7 +1,0 @@
-function createPerson(name) {
-    return { name: name };
-}
-var myPerson1 = createPerson();
-var myPerson2 = createPerson("Robson");
-console.log(myPerson1);
-console.log(myPerson2);
